+++ tests/Datasets/LabeledTest.php
@@ -249,12 +249,24 @@ class LabeledTest extends TestCase
         $this->assertCount(1, $this->dataset);
     }
 
+    public function test_slice_dataset()
+    {
+        $this->assertCount(6, $this->dataset);
+
+        $subset = $this->dataset->slice(2, 2);
+
+        $this->assertInstanceOf(Labeled::class, $subset);
+        $this->assertCount(2, $subset);
+        $this->assertCount(6, $this->dataset);
+    }
+
     public function test_splice_dataset()
     {
         $this->assertCount(6, $this->dataset);
 
         $subset = $this->dataset->splice(2, 2);
 
+        $this->assertInstanceOf(Labeled::class, $subset);
         $this->assertCount(2, $subset);
         $this->assertCount(4, $this->dataset);
     }